x86:asm:processor: Verify if CPU provides LASS support
The Intel SRF (Sierra-Forest) CPUs support the Linear Address Space
Separation (LASS) feature, which enforces paging mode-based protection
and returns the GPF fault in the situation, when the software tries
to access memory space addresses, protected by LASS.

The LASS feature support can be verified by the usage of the CPUID
instruction, returning the Enumeration Sub-leaf (EAX = 07H and
ECX = 01H). After it's execution CPU should set bit 6 in the EAX,
in case when the LASS is supported.

Link: Intel SDM - https://cdrdv2.intel.com/v1/dl/getContent/67120
Signed-off-by: Michal Camacho Romero <michal.camacho.romero@intel.com>